refactor: InputView와 ResultView의 기능을 명확히 함
