+++ src/main/java/lotto/view/InputView.java
@@ -1,11 +1,8 @@
 package lotto.view;
 
-import java.util.List;
 import java.util.Scanner;
-import java.util.stream.Collectors;
 
 import lotto.domain.LottoNumber;
-import lotto.domain.LottoTicket;
 import lotto.domain.LottoTickets;
 import lotto.domain.WinningLottoTicket;
 
@@ -21,29 +18,7 @@ public class InputView {
 		System.out.println("구입금액을 입력해 주세요.");
 		int cash = scanner.nextInt();
 		LottoTickets lottoTickets = new LottoTickets(cash);
-
-		printTicketCount(lottoTickets.getTicketCount());
-		printTickets(lottoTickets.getLottoTickets());
-
 		return lottoTickets;
-	}
-
-	private static void printTicketCount(int ticketCount) {
-		System.out.println(ticketCount + "개를 구매했습니다.");
-	}
-
-	private static void printTickets(List<LottoTicket> lottoTickets) {
-		for (LottoTicket lottoTicket : lottoTickets) {
-			printLottoTicketNumbers(lottoTicket);
-		}
-	}
-
-	private static void printLottoTicketNumbers(LottoTicket lottoTicket) {
-		List<Integer> numbers = lottoTicket.get().stream()
-			.map(lottoNumber -> lottoNumber.getNumber())
-			.collect(Collectors.toList());
-
-		System.out.println(numbers.toString());
 	}
 
 	public static WinningLottoTicket requireWinningNumber() {
